+++ consensus/doc/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.20] - 2023-01-19
+
+- Rewrite `interface{}` to `any`
+
+## [0.0.0.19] - 2023-01-18
+
+- Remove `Block` proto definition to consolidate under `shared/core/types`
+
 ## [0.0.0.18] - 2023-01-11
 
 ### Consensus - Core