+++ utility/doc/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.20] - 2023-01-20
+
+- Remove `address []byte` argument from `InsertPool` function
+
+## [0.0.0.19] - 2023-01-19
+
+- Rewrite `interface{}` to `any`
+
+## [0.0.0.18] - 2023-01-14
+
+- Updated "test/module_test.go" to encorporate MaxConnsCount, MinConnsCount, MaxConnLifetime, MaxConnIdleTime, and HealthCheckPeriod in persistence config.
+
 ## [0.0.0.17] - 2023-01-11
 
 - Deprecated `GetBlocksPerSession()` in favour of the more general parameter getter function `GetParameter()`